Move tests command line interface to a separate file
