+++ .travis.yml
@@ -19,4 +19,4 @@ install:
 
 # command to run tests
 script:
-  - ./dakara.py test
+  - ./tests.py
